+++ AudioAnalyzer/Sources/audio-utils/CustomizableValueTransformer.h
@@ -63,7 +63,7 @@ namespace rxtd::audio_utils {
 		}
 
 		[[nodiscard]]
-		float apply(double value);
+		double apply(double value);
 
 		[[nodiscard]]
 		float apply(float value) {
